Toggle recording/playback with spacebar
This is implemented by switching keyboard focus to control for the previous
state on activation.

The spacebar shortcut previously associated with the recording control in the
toolbar has been removed so it can return to its default behavior of activating
whichever control is currently focused. This also means that tab can be used
to move focus amongst playback controls and activate them.